chore: fix dependency issue
