+++ src/components/PriceAlert/PriceAlert.tsx
@@ -2,7 +2,7 @@ import { FormattedMessage } from '@edx/frontend-platform/i18n';
 import { Alert } from '@openedx/paragon';
 
 import { DisplayPrice } from '@/components/DisplayPrice';
-import { SUBSCRIPTION_PRICE_PER_USER_PER_MONTH } from '@/components/Stepper/constants';
+import { SUBSCRIPTION_PRICE_PER_USER_PER_MONTH } from '@/constants/checkout';
 
 const PriceAlert = () => (
   <Alert className="price-alert-background">
