+++ src/components/Stepper/index.ts
@@ -1,3 +1,2 @@
 export { default as CheckoutStepper } from './CheckoutStepper';
-export * from './constants';
 export * from './Steps';
