+++ src/components/billing-details-pages/BillingDetailsPage.tsx
@@ -10,11 +10,10 @@ import { Helmet } from 'react-helmet';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 
-import { DataStoreKey } from '@/components/Stepper/constants';
 import { useStepperContent } from '@/components/Stepper/Steps/hooks';
 import {
   CheckoutPageDetails,
-  CheckoutStepKey,
+  CheckoutStepKey, DataStoreKey,
 } from '@/constants/checkout';
 import { useCheckoutFormStore, useCurrentPageDetails } from '@/hooks/index';
